+++ src/aiconfigurator/generator/rendering/schemas.py
@@ -8,6 +8,7 @@ from typing import Any, Optional
 
 import yaml
 
+from ..utils import get_default_dynamo_version_mapping
 from .engine import evaluate_expression
 
 _SCHEMA_CACHE: dict[str, list[dict[str, Any]]] = {}
@@ -15,8 +16,6 @@ _DEFAULT_BACKEND = "trtllm"
 _BASE_DIR = Path(__file__).resolve().parent
 _CONFIG_DIR = _BASE_DIR.parent / "config"
 _SCHEMA_FILE = (_CONFIG_DIR / "deployment_config.yaml").resolve()
-_BACKEND_VERSION_MATRIX_FILE = (_CONFIG_DIR / "backend_version_matrix.yaml").resolve()
-_MATRIX_CACHE: dict[str, dict[str, Any]] = {}
 
 
 def _load_schema_inputs(schema_path: str) -> list[dict[str, Any]]:
@@ -36,30 +35,6 @@ def _load_schema_inputs(schema_path: str) -> list[dict[str, Any]]:
         inputs = []
     _SCHEMA_CACHE[path] = inputs
     return inputs
-
-
-def _load_backend_version_matrix(matrix_path: str) -> dict[str, dict[str, Any]]:
-    path = os.path.abspath(str(matrix_path))
-    cached = _MATRIX_CACHE.get(path)
-    if cached is not None:
-        return cached
-    try:
-        with open(path, encoding="utf-8") as f:
-            payload = yaml.safe_load(f) or {}
-    except Exception:
-        payload = {}
-    matrix = payload.get("matrix", payload) if isinstance(payload, dict) else {}
-    if not isinstance(matrix, dict):
-        matrix = {}
-    _MATRIX_CACHE[path] = matrix
-    return matrix
-
-
-def get_default_dynamo_version() -> Optional[str]:
-    matrix = _load_backend_version_matrix(_BACKEND_VERSION_MATRIX_FILE)
-    if not matrix:
-        return None
-    return str(next(iter(matrix.keys())))
 
 
 def _normalize_backend(backend: Optional[str]) -> str:
@@ -101,7 +76,7 @@ def apply_defaults(
     if extra_context:
         eval_ctx.update(extra_context)
     if group == "K8sConfig" and not eval_ctx.get("generator_dynamo_version"):
-        default_version = get_default_dynamo_version()
+        default_version, _ = get_default_dynamo_version_mapping()
         if default_version:
             eval_ctx["generator_dynamo_version"] = default_version
     out = dict(cfg)
